requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0

# LLM Integrations
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.7.0

# File Parsing
pypdf>=4.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
pytesseract>=0.3.10
pypdfium2>=4.30.0

# Visualization & Charts
matplotlib>=3.8.0
plotly>=5.17.0
seaborn>=0.13.0

# PDF Generation
reportlab>=4.0.0

# Web Framework
streamlit>=1.28.0
streamlit-sortables>=0.2.0
requests>=2.31.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.8.0

# Optional: For advanced features
pillow>=10.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
